requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
protobuf==2.6.1
pycrypto==2.6.1
requests==2.6.0
Werkzeug==0.10.4
redis==2.10.3
